requirements.txt
discord.py[voice]
wavelink
flask
requests
python-dotenv
